back/pyproject.toml
[build-system]
requires = ["poetry==0.12.8"]
build-backend = "poetry.masonry.api"


[tool.poetry]
name = "euroet"
description = "Euroet engineering tools"
version = "0.2.0"
readme = "README.md"
authors = []


[tool.poetry.dependencies]
python = "3.6.8"
django = "^2.1"
django-split-settings = "^0.3"
django-axes = "^4.5"
django-csp = "^3.5"
django-referrer-policy = "^1.0"
django-health-check = "^3.9"
gunicorn = "^19.9"
python-decouple = "^3.1"
dump-env = "^0.2"
bcrypt = "^3.1"
psycopg2-binary = "^2.7"
djangorestframework = "^3.9"
django-rest-knox = "^3.4"
colorhash = "^1.0"
django-modeladmin-reorder = "^0.3.1"
coreapi = "^2.3"
pyparsing = "^2.4"
XlsxWriter = "^1.1"
django-constance = {version = "^2.4",extras = ["database"]}
django-ordered-model = "^3.3"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^1.11"
django-querycount = "^0.7"
nplusone = "^1.0"
wemake-python-styleguide = "^0.7"
flake8-pytest = "^1.3"
flake8-django = "^0.0.4"
pytest = "^4.3"
pytest-flake8 = "^1.0"
pytest-isort = "^0.2"
pytest-django = "^3.4"
pytest-cov = "^2.6"
pytest-randomly = "^1.2"
pytest-deadfixtures = "^2.1"
pytest-testmon = "^0.9"
pytest-timeout = "^1.3"
pre-commit = "^1.14"
xenon = "^0.5"
mypy = "^0.670"
ipython = "^7.3"
sphinx = "^1.8"
sphinx-autodoc-typehints = "^1.6"
doc8 = "^0.8"
yamllint = "^1.15"
safety = "^1.8"
dotenv-linter = "^0.1"
polint = "^0.4"
black = {version = "^18.3-alpha.0",allows-prereleases = true}
django-extensions = "^2.1"
django-webpack-loader = "^0.6.0"
isort = "^4.3"
ipdb = "^0.11.0"
faker = "^1.0"
factory_boy = "^2.11"
readline = "^6.2"


[tool.black]
py36 = true
line-length = 118
skip-string-normalization = true
exclude = ".git|.mypy_cache|front|migrations/|settings/|urls.py"
